+++ Sample.Domain/SampleEntityTypes.cs
@@ -13,6 +13,5 @@ public enum SampleEntityTypes : byte
     QuestionComment = 5,
     Tag = 6,
     User = 7,
-    Author = 8,
-    TestEntity = 255
+    Author = 8
 }